decorator ConvertToStringWithIndent now accepts strings
util.CreateStringFromList now checks if the argument is a string, and if so,
handles it without adding on braces at either end.

This allows the ToStr functions to return plain strings as well as lists.
Returning a list means is kind of equivalent to returning an s-expression,
and returning a string is kind of equivalent to just returning an atom
used currently in the VariableReference, etc..